[EMBR-1196] Move FailedApiCallsPerEndpoint to another file.
